# feat: concurrent file downloads for localization