--- a/wdl-engine/tests/workflows/concurrent-localization/local.txt
+++ b/wdl-engine/tests/workflows/concurrent-localization/local.txt
@@ -1,0 +1,1 @@
+This is the local test file.