--- a/wdl-engine/tests/workflows/concurrent-localization/outputs.json
+++ b/wdl-engine/tests/workflows/concurrent-localization/outputs.json
@@ -1,0 +1,303 @@
+{
+  "test.out1": {
+    "userId": 1,
+    "id": 1,
+    "title": "delectus aut autem",
+    "completed": false
+  },
+  "test.out2": {
+    "userId": 1,
+    "id": 2,
+    "title": "quis ut nam facilis et officia qui",
+    "completed": false
+  },
+  "test.out3": {
+    "userId": 1,
+    "id": 3,
+    "title": "fugiat veniam minus",
+    "completed": false
+  },
+  "test.out4": {
+    "userId": 1,
+    "id": 4,
+    "title": "et porro tempora",
+    "completed": true
+  },
+  "test.out5": {
+    "userId": 1,
+    "id": 5,
+    "title": "laboriosam mollitia et enim quasi adipisci quia provident illum",
+    "completed": false
+  },
+  "test.out6": {
+    "userId": 1,
+    "id": 6,
+    "title": "qui ullam ratione quibusdam voluptatem quia omnis",
+    "completed": false
+  },
+  "test.out7": {
+    "userId": 1,
+    "id": 7,
+    "title": "illo expedita consequatur quia in",
+    "completed": false
+  },
+  "test.out8": {
+    "userId": 1,
+    "id": 8,
+    "title": "quo adipisci enim quam ut ab",
+    "completed": true
+  },
+  "test.out9": {
+    "userId": 1,
+    "id": 9,
+    "title": "molestiae perspiciatis ipsa",
+    "completed": false
+  },
+  "test.out10": {
+    "userId": 1,
+    "id": 10,
+    "title": "illo est ratione doloremque quia maiores aut",
+    "completed": true
+  },
+  "test.out11": {
+    "userId": 1,
+    "id": 11,
+    "title": "vero rerum temporibus dolor",
+    "completed": true
+  },
+  "test.out12": {
+    "userId": 1,
+    "id": 12,
+    "title": "ipsa repellendus fugit nisi",
+    "completed": true
+  },
+  "test.out13": {
+    "userId": 1,
+    "id": 13,
+    "title": "et doloremque nulla",
+    "completed": false
+  },
+  "test.out14": {
+    "userId": 1,
+    "id": 14,
+    "title": "repellendus sunt dolores architecto voluptatum",
+    "completed": true
+  },
+  "test.out15": {
+    "userId": 1,
+    "id": 15,
+    "title": "ab voluptatum amet voluptas",
+    "completed": true
+  },
+  "test.out16": {
+    "userId": 1,
+    "id": 16,
+    "title": "accusamus eos facilis sint et aut voluptatem",
+    "completed": true
+  },
+  "test.out17": {
+    "userId": 1,
+    "id": 17,
+    "title": "quo laboriosam deleniti aut qui",
+    "completed": true
+  },
+  "test.out18": {
+    "userId": 1,
+    "id": 18,
+    "title": "dolorum est consequatur ea mollitia in culpa",
+    "completed": false
+  },
+  "test.out19": {
+    "userId": 1,
+    "id": 19,
+    "title": "molestiae ipsa aut voluptatibus pariatur dolor nihil",
+    "completed": true
+  },
+  "test.out20": {
+    "userId": 1,
+    "id": 20,
+    "title": "ullam nobis libero sapiente ad optio sint",
+    "completed": true
+  },
+  "test.out21": {
+    "userId": 2,
+    "id": 21,
+    "title": "suscipit repellat esse quibusdam voluptatem incidunt",
+    "completed": false
+  },
+  "test.out22": {
+    "userId": 2,
+    "id": 22,
+    "title": "distinctio vitae autem nihil ut molestias quo",
+    "completed": true
+  },
+  "test.out23": {
+    "userId": 2,
+    "id": 23,
+    "title": "et itaque necessitatibus maxime molestiae qui quas velit",
+    "completed": false
+  },
+  "test.out24": {
+    "userId": 2,
+    "id": 24,
+    "title": "adipisci non ad dicta qui amet quaerat doloribus ea",
+    "completed": false
+  },
+  "test.out25": {
+    "userId": 2,
+    "id": 25,
+    "title": "voluptas quo tenetur perspiciatis explicabo natus",
+    "completed": true
+  },
+  "test.out26": {
+    "userId": 2,
+    "id": 26,
+    "title": "aliquam aut quasi",
+    "completed": true
+  },
+  "test.out27": {
+    "userId": 2,
+    "id": 27,
+    "title": "veritatis pariatur delectus",
+    "completed": true
+  },
+  "test.out28": {
+    "userId": 2,
+    "id": 28,
+    "title": "nesciunt totam sit blanditiis sit",
+    "completed": false
+  },
+  "test.out29": {
+    "userId": 2,
+    "id": 29,
+    "title": "laborum aut in quam",
+    "completed": false
+  },
+  "test.out30": {
+    "userId": 2,
+    "id": 30,
+    "title": "nemo perspiciatis repellat ut dolor libero commodi blanditiis omnis",
+    "completed": true
+  },
+  "test.out31": {
+    "userId": 2,
+    "id": 31,
+    "title": "repudiandae totam in est sint facere fuga",
+    "completed": false
+  },
+  "test.out32": {
+    "userId": 2,
+    "id": 32,
+    "title": "earum doloribus ea doloremque quis",
+    "completed": false
+  },
+  "test.out33": {
+    "userId": 2,
+    "id": 33,
+    "title": "sint sit aut vero",
+    "completed": false
+  },
+  "test.out34": {
+    "userId": 2,
+    "id": 34,
+    "title": "porro aut necessitatibus eaque distinctio",
+    "completed": false
+  },
+  "test.out35": {
+    "userId": 2,
+    "id": 35,
+    "title": "repellendus veritatis molestias dicta incidunt",
+    "completed": true
+  },
+  "test.out36": {
+    "userId": 2,
+    "id": 36,
+    "title": "excepturi deleniti adipisci voluptatem et neque optio illum ad",
+    "completed": true
+  },
+  "test.out37": {
+    "userId": 2,
+    "id": 37,
+    "title": "sunt cum tempora",
+    "completed": false
+  },
+  "test.out38": {
+    "userId": 2,
+    "id": 38,
+    "title": "totam quia non",
+    "completed": false
+  },
+  "test.out39": {
+    "userId": 2,
+    "id": 39,
+    "title": "doloremque quibusdam asperiores libero corrupti illum qui omnis",
+    "completed": false
+  },
+  "test.out40": {
+    "userId": 2,
+    "id": 40,
+    "title": "totam atque quo nesciunt",
+    "completed": true
+  },
+  "test.out41": {
+    "userId": 3,
+    "id": 41,
+    "title": "aliquid amet impedit consequatur aspernatur placeat eaque fugiat suscipit",
+    "completed": false
+  },
+  "test.out42": {
+    "userId": 3,
+    "id": 42,
+    "title": "rerum perferendis error quia ut eveniet",
+    "completed": false
+  },
+  "test.out43": {
+    "userId": 3,
+    "id": 43,
+    "title": "tempore ut sint quis recusandae",
+    "completed": true
+  },
+  "test.out44": {
+    "userId": 3,
+    "id": 44,
+    "title": "cum debitis quis accusamus doloremque ipsa natus sapiente omnis",
+    "completed": true
+  },
+  "test.out45": {
+    "userId": 3,
+    "id": 45,
+    "title": "velit soluta adipisci molestias reiciendis harum",
+    "completed": false
+  },
+  "test.out46": {
+    "userId": 3,
+    "id": 46,
+    "title": "vel voluptatem repellat nihil placeat corporis",
+    "completed": false
+  },
+  "test.out47": {
+    "userId": 3,
+    "id": 47,
+    "title": "nam qui rerum fugiat accusamus",
+    "completed": false
+  },
+  "test.out48": {
+    "userId": 3,
+    "id": 48,
+    "title": "sit reprehenderit omnis quia",
+    "completed": false
+  },
+  "test.out49": {
+    "userId": 3,
+    "id": 49,
+    "title": "ut necessitatibus aut maiores debitis officia blanditiis velit et",
+    "completed": false
+  },
+  "test.out50": {
+    "userId": 3,
+    "id": 50,
+    "title": "cupiditate necessitatibus ullam aut quis dolor voluptate",
+    "completed": true
+  },
+  "test.relative_out": "This is the local test file."
+}