@dead8309 2025-04-18
test runs with 50 remote files

#### Before

- Uncached

```bash
➜  wdl git:(feat/concurrent-downloads) time cargo test --package wdl-engine --features="codespan" --test workflows -- -- concurrent-localization
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.19s
     Running tests/workflows.rs (target/debug/deps/workflows-719bfd71ea48623c)

running 1 tests

test concurrent-localization ... ok

test result: ok. 1 passed

cargo test --package wdl-engine --features="codespan" --test workflows -- --   0.86s user 0.25s system 15% cpu 7.383 total
```

- Cached

```bash
➜  wdl git:(feat/concurrent-downloads) time cargo test --package wdl-engine --features="codespan" --test workflows -- -- concurrent-localization
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.20s
     Running tests/workflows.rs (target/debug/deps/workflows-719bfd71ea48623c)

running 1 tests

test concurrent-localization ... ok

test result: ok. 1 passed

cargo test --package wdl-engine --features="codespan" --test workflows -- --   0.62s user 0.16s system 35% cpu 2.196 total
```

---

### After

- Uncached
```bash

➜  wdl git:(feat/concurrent-downloads) ✗ time cargo test --package wdl-engine --features="codespan" --test workflows -- -- concurrent-localization
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.18s
     Running tests/workflows.rs (target/debug/deps/workflows-719bfd71ea48623c)

running 1 tests

test concurrent-localization ... ok

test result: ok. 1 passed

cargo test --package wdl-engine --features="codespan" --test workflows -- --   0.93s user 0.31s system 38% cpu 3.265 total
```

- Cached

```bash
➜  wdl git:(feat/concurrent-downloads) ✗ time cargo test --package wdl-engine --features="codespan" --test workflows -- -- concurrent-localization
    Finished `test` profile [unoptimized + debuginfo] target(s) in 0.18s
     Running tests/workflows.rs (target/debug/deps/workflows-719bfd71ea48623c)

running 1 tests

test concurrent-localization ... ok

test result: ok. 1 passed

cargo test --package wdl-engine --features="codespan" --test workflows -- --   0.67s user 0.16s system 45% cpu 1.832 total
```