@dead8309 2025-04-18
tldr;
| Type | Uncached Time | Cached Time |
| -----|----------------|--------------|
| Before | 7.383s | 2.196s |
| After | 3.265s | 1.832s |