@claymcleod 2025-04-19
With this change, I think it's also prudent to go ahead and handle some rudimentary concept of retries within the `Downloader`. This is mainly because I think it's more likely that these are going to fail due to network bandwidth issues. My suggestion is to acquire the permit _within_ the loop (getting a new permit on each retry) so that you fairly share downloading bandwidth amongst the files (and, perhaps give the particular files a second to breathe if the download has just failed. It's unclear how often that will make a difference, but it can't hurt I would say).